Update README.md
The script needs to run continuously to perform the desired task.

To run the code continuously into a Raspberry Pi device, you need to add the script into Pi's boot menu to automatically start running when powered on. Follow the below methods to do the same:
1. https://www.codementor.io/@gergelykovcs/how-to-run-and-schedule-python-scripts-on-raspberry-pi-n2clhe3kp
2. https://www.youtube.com/watch?v=cZa1oCSdbUk&t=16s
3. https://www.youtube.com/watch?v=zRXauWUumSI (Easiest Way)


After uploading the script into Pi just relax and enjoy the effortless login from now on.